+++ Automate/Framework/Machines/Objects/OstrichIncubatorMachine.cs
@@ -25,7 +25,7 @@ namespace Pathoschild.Stardew.Automate.Framework.Machines.Objects
         public OstrichIncubatorMachine(SObject machine, GameLocation location, Vector2 tile)
             : base(machine, location, tile)
         {
-            int minutesUntilReady = Game1.player.professions.Contains(Farmer.butcher) ? 7500 : 15000; // coopmaster
+            int minutesUntilReady = this.GetOwner().professions.Contains(Farmer.butcher) ? 7500 : 15000; // coopmaster
             this.Recipes = new IRecipe[]
             {
                 // ostrich egg => ostrich